removed Comment folder and moved components to Common folder
