+++ src/pages/c/[community]/[postId]/[slug]/index.tsx
@@ -21,8 +21,8 @@ import { createContext } from "src/backend/context";
 import { createSSGHelpers } from "@trpc/react/ssg";
 import { trpc } from "@/utils/trpc";
 import Markdown from "@/components/common/Markdown";
-import Comment from "@/components/Comments/Comment";
-import CommentSkeleton from "@/components/Comments/CommentSkeleton";
+import Comment from "@/components/common/Comment";
+import CommentSkeleton from "@/components/common/CommentSkeleton";
 import MarkdownTipsModal from "@/components/common/MarkdownTipsModal";
 import TextareaAutosize from "@/components/common/TextareaAutosize";
